Add files via upload
Windows Activation Tool

This batch script is a tool that can be used to activate Windows 11, 10, 8.1, 8, and 7. It is a simple and easy-to-use tool that can be used by anyone, regardless of their technical expertise.

Features:

Activates Windows 11, 10, 8.1, 8, and 7
Simple and easy to use
Does not require any technical expertise
Safe and secure
How to use:

Download the batch script from Github.